移除 std::ranges::sample 参数的类型强制转换
将 std::ranges::sample 的第三个参数由 static_cast<int>(unool::math::floor(hand.count() / 3.0)) 改为直接使用 unool::math::floor(hand.count() / 3.0)，简化了代码并去除了不必要的类型转换。

+++ source/ASkill.cpp
@@ -17,7 +17,7 @@ bool 装弹::content(GameLogic& game, Player& carrier) {
 	std::ranges::sample(
 		hand | std::views::transform([](const auto& p) { return std::ref(*p); }),
 		std::back_inserter(cards),
-		static_cast<int>(unool::math::floor(hand.count() / 3.0)),
+		unool::math::floor(hand.count() / 3.0),
 		unool::random::rng
 	);
 	for (Card& c : cards) {
